requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.21
psycopg[binary,pool]==3.2.1
pydantic==2.9.2
numpy==2.1.1
pandas==2.2.2
pgvector==0.3.3
python-dotenv==1.0.1
markdown-it-py==3.0.0
pypdf==4.3.1
